cleanup: separate out roam utils
