+++ formats/roam/roam.ts
@@ -2,7 +2,8 @@ import { ImportResult } from '../../main';
 import { RoamPage, RoamBlock, JsonObject, BlockInfo } from './models/roam-json';
 import { PickedFile, path, fs } from 'filesystem';
 import { RoamJSONImporter } from 'formats/roam-json';
-import { sanitizeFileName, sanitizeFileNameKeepPath, getUserDNPFormat, convertDateString } from '../../util';
+import { sanitizeFileName} from '../../util';
+import { sanitizeFileNameKeepPath, getUserDNPFormat, convertDateString  } from './roam_utils';
 import { TFile, TFolder } from 'obsidian';
 import { downloadFirebaseFile } from './roam_dl_attachment';
 import { ProgressReporter } from '../../main';
